修正内容反映1

+++ app/controllers/tweets_controller.rb
@@ -2,22 +2,13 @@ class TweetsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :create]
 
   def index
-    @tweets = Tweet.all
+  @tweets = Tweet.all
 
-    if params[:tag_ids]
-      @tweets = []
-      params[:tag_ids].each do |key, value|
-        if value == "1"
-          tag_tweets = Tag.find_by(name: key).tweets
-          @tweets = @tweets.empty? ? tag_tweets : @tweets & tag_tweets
-        end
-      end
-    end
-
-    if params[:tag]
-      Tag.create(name: params[:tag])
-    end
+  if params[:tag_ids].present?
+    @tweets = @tweets.joins(:tags).where(tags: { id: params[:tag_ids] }).distinct
   end
+end
+
 
   def new
     @tweet = Tweet.new
@@ -34,20 +25,19 @@ class TweetsController < ApplicationController
   end
 
   def show
-        @tweet = Tweet.find(params[:id])
-        @comments = @tweet.comments
-        @comment = Comment.new
+    @tweet = Tweet.find(params[:id])
+    @comments = @tweet.comments
+    @comment = Comment.new
   end
 
-def destroy
-   tweet = Tweet.find(params[:id])
-   if tweet.user_id == current_user.id
-    tweet.destroy
-   end
-   redirect_to action: "index"
-end
-   
+  def destroy
+    tweet = Tweet.find(params[:id])
+    tweet.destroy if tweet.user_id == current_user.id
+    redirect_to action: "index"
+  end
+
   private
+
   def tweet_params
     params.require(:tweet).permit(:genre, :purpose, :place, :youtube_url, :body, :video, tag_ids: [])
   end
